requirements.txt
# Основные зависимости FastAPI
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic[email]==2.5.0
python-multipart==0.0.6

# Google OAuth зависимости
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
authlib==1.2.1

# JWT и криптография
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4

# HTTP клиент
httpx==0.25.2
requests==2.31.0

# Middleware
starlette==0.27.0
itsdangerous==2.1.2

# WhisperX и ML зависимости
whisperx
torch
torchaudio
transformers
accelerate

# Обработка файлов
python-docx==1.1.0
reportlab==4.0.7

# AWS/S3
boto3==1.34.0

# Real-time зависимости
numpy>=1.21.0
websockets>=11.0.0
scipy>=1.9.0

# Утилиты
python-dotenv==1.0.0 